Move model/update/action/env into SubconsciousApp
