--- xcode/Subconscious/Shared/Services/AppEnvironment.swift
@@ -1,153 +1,0 @@
-//
-//  AppEnvironment.swift
-//  Subconscious
-//
-//  Created by Gordon Brander on 9/16/21.
-//
-
-import Foundation
-import Combine
-import os
-
-/// A place for constants and services
-struct AppEnvironment {
-    var documentURL: URL
-    var applicationSupportURL: URL
-
-    var logger: Logger
-    var keyboard: KeyboardService
-    var database: DatabaseService
-    /// Issues periodic polls
-
-    /// Create a long polling publisher that never completes
-    static func poll(every interval: Double) -> AnyPublisher<Date, Never> {
-        Timer.publish(
-            every: interval,
-            on: .main,
-            in: .default
-        )
-        .autoconnect()
-        .eraseToAnyPublisher()
-    }
-
-    init() {
-        self.documentURL = FileManager.default.urls(
-            for: .documentDirectory,
-               in: .userDomainMask
-        ).first!
-
-        self.applicationSupportURL = try! FileManager.default.url(
-            for: .applicationSupportDirectory,
-            in: .userDomainMask,
-            appropriateFor: nil,
-            create: true
-        )
-
-        self.logger = Logger(
-            subsystem: Config.rdns,
-            category: "main"
-        )
-
-        self.database = DatabaseService(
-            documentURL: self.documentURL,
-            databaseURL: self.applicationSupportURL
-                .appendingPathComponent("database.sqlite"),
-            migrations: Self.migrations
-        )
-
-        self.keyboard = KeyboardService()
-    }
-}
-
-//  MARK: Migrations
-extension AppEnvironment {
-    static let migrations = SQLite3Migrations([
-        SQLite3Migrations.Migration(
-            date: "2021-11-04T12:00:00",
-            sql: """
-            CREATE TABLE search_history (
-                id TEXT PRIMARY KEY,
-                query TEXT NOT NULL,
-                created TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-            );
-
-            CREATE TABLE entry (
-              slug TEXT PRIMARY KEY,
-              title TEXT NOT NULL DEFAULT '',
-              body TEXT NOT NULL,
-              modified TEXT NOT NULL,
-              size INTEGER NOT NULL
-            );
-
-            CREATE VIRTUAL TABLE entry_search USING fts5(
-              slug,
-              title,
-              body,
-              modified UNINDEXED,
-              size UNINDEXED,
-              content="entry",
-              tokenize="porter"
-            );
-
-            /*
-            Create triggers to keep fts5 virtual table in sync with content table.
-
-            Note: SQLite documentation notes that you want to modify the fts table *before*
-            the external content table, hence the BEFORE commands.
-
-            These triggers are adapted from examples in the docs:
-            https://www.sqlite.org/fts3.html#_external_content_fts4_tables_
-            */
-            CREATE TRIGGER entry_search_before_update BEFORE UPDATE ON entry BEGIN
-              DELETE FROM entry_search WHERE rowid=old.rowid;
-            END;
-
-            CREATE TRIGGER entry_search_before_delete BEFORE DELETE ON entry BEGIN
-              DELETE FROM entry_search WHERE rowid=old.rowid;
-            END;
-
-            CREATE TRIGGER entry_search_after_update AFTER UPDATE ON entry BEGIN
-              INSERT INTO entry_search
-                (
-                  rowid,
-                  slug,
-                  title,
-                  body,
-                  modified,
-                  size
-                )
-              VALUES
-                (
-                  new.rowid,
-                  new.slug,
-                  new.title,
-                  new.body,
-                  new.modified,
-                  new.size
-                );
-            END;
-
-            CREATE TRIGGER entry_search_after_insert AFTER INSERT ON entry BEGIN
-              INSERT INTO entry_search
-                (
-                  rowid,
-                  slug,
-                  title,
-                  body,
-                  modified,
-                  size
-                )
-              VALUES
-                (
-                  new.rowid,
-                  new.slug,
-                  new.title,
-                  new.body,
-                  new.modified,
-                  new.size
-                );
-            END;
-            """
-        )!
-    ])!
-}
